# org.oasis.oslcop.sysml.oslc-server/src/main/resources/swagger (1).yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: ""
host: "jamsden.rtp.raleigh.ibm.com:8443"
basePath: "/sysml"
...
schemes:
- "https"
paths:
...
  /crud/projects/{serviceProviderId}/Element/{id}:
    get:
      tags:
      - "Web Service for {Element, Requirement, Relationship}"
      summary: "GET for resources of type {'Element'}"
      description: "GET for resources of type {'<a href=\"http://oasis.oslc-op/ns/sysml#Element\"\
        >Element</a>'}, with respective resource shapes {'<a href=\"../services/resourceShapes/element\"\
        >Element</a>'}"
      operationId: "getElementCompact"
      produces:
      - "application/rdf+xml"
      - "application/xml"
      - "application/json"
      - "text/turtle"
      - "text/html"
      - "application/x-oslc-compact+xml"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Compact"
    put:
      tags:
      - "Web Service for {Element, Requirement, Relationship}"
      summary: "PUT for resources of type {'Element'}"
      description: "PUT for resources of type {'<a href=\"http://oasis.oslc-op/ns/sysml#Element\"\
        >Element</a>'}, with respective resource shapes {'<a href=\"../services/resourceShapes/element\"\
        >Element</a>'}"
      operationId: "updateElement"
      consumes:
      - "application/rdf+xml"
      - "application/ld+json"
      - "text/turtle"
      - "application/xml"
      - "application/json"
      produces:
      - "application/rdf+xml"
      - "application/xml"
      - "application/json"
      - "text/turtle"
      - "text/html"
      - "application/x-oslc-compact+xml"
      parameters:
      - name: "If-Match"
        in: "header"
        required: false
        type: "string"
      - name: "serviceProviderId"
        in: "path"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Element"
      responses:
        default:
          description: "successful operation"
    delete:
      tags:
      - "Web Service for {Element, Requirement, Relationship}"
      summary: "DELETE for resources of type {'Element'}"
      description: "DELETE for resources of type {'<a href=\"http://oasis.oslc-op/ns/sysml#Element\"\
        >Element</a>'}, with respective resource shapes {'<a href=\"../services/resourceShapes/element\"\
        >Element</a>'}"
      operationId: "deleteElement"
      produces:
      - "application/rdf+xml"
      - "application/xml"
      - "application/json"
      - "text/turtle"
      - "text/html"
      - "application/x-oslc-compact+xml"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "successful operation"
  /crud/projects/{serviceProviderId}/Element/{id}/smallPreview:
    get:
      tags:
      - "Web Service for {Element, Requirement, Relationship}"
      operationId: "getElementAsHtmlSmallPreview"
      produces:
      - "text/html"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "successful operation"
  /crud/projects/{serviceProviderId}/Element/{id}/largePreview:
    get:
      tags:
      - "Web Service for {Element, Requirement, Relationship}"
      operationId: "getElementAsHtmlLargePreview"
      produces:
      - "text/html"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        default:
          description: "successful operation"
...
definitions:
...
  Element:
    type: "object"
    properties:
      about:
        type: "string"
        format: "uri"
      types:
        type: "array"
        items:
          type: "string"
          format: "uri"
      extendedProperties:
        type: "object"
        additionalProperties:
          type: "object"
      contributor:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      created:
        type: "string"
        format: "date-time"
      creator:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      description:
        type: "string"
      identifier:
        type: "string"
      modified:
        type: "string"
        format: "date-time"
      source:
        type: "string"
        format: "uri"
      title:
        type: "string"
      type:
        type: "array"
        uniqueItems: true
        items:
          type: "string"
      instanceShape:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      serviceProvider:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      shortTitle:
        type: "string"
      external:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      trace:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      refine:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      derives:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      elaborates:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      satisfy:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
      ownedRelationship:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Link"
  Link:
    type: "object"
    properties:
      value:
        type: "string"
        format: "uri"
      label:
        type: "string"
  Compact:
    type: "object"
    properties:
      about:
        type: "string"
        format: "uri"
      types:
        type: "array"
        items:
          type: "string"
          format: "uri"
      extendedProperties:
        type: "object"
        additionalProperties:
          type: "object"
      icon:
        type: "string"
        format: "uri"
      largePreview:
        $ref: "#/definitions/Preview"
      shortTitle:
        type: "string"
      smallPreview:
        $ref: "#/definitions/Preview"
      title:
        type: "string"
  Preview:
    type: "object"
    properties:
      about:
        type: "string"
        format: "uri"
      types:
        type: "array"
        items:
          type: "string"
          format: "uri"
      extendedProperties:
        type: "object"
        additionalProperties:
          type: "object"
      document:
        type: "string"
        format: "uri"
      hintHeight:
        type: "string"
      hintWidth:
        type: "string"
      initialHeight:
        type: "string"
...